separated into rails folder
