+++ lib/generators/rails/templates/migration.rb.erb
